requirements.txt
TgCrypto
asyncio
ffmpeg-python
py-tgcalls
git+https://github.com/pyrogram/pyrogram@master
youtube-search-python==1.4.9
youtube-search
httpx==0.13.3
python-dotenv
dnspython
gitpython
aiofiles
aiohttp
requests
heroku3
motor
yt-dlp
psutil
future
wget
